Crate mpot_gen

Modules

anyhow

github crates-io docs-rs

cal 🔒

log

A lightweight logging facade.

mail

Types for processing new posts: PostFilter, ListContext, MailJob and PostAction.

melib

A crate that performs mail client operations such as

models

Database models: MailingList, ListOwner, ListSubscription, PostPolicy, SubscriptionPolicy and Post.

postfix

Generate configuration for the postfix mail server.

rusqlite

Rusqlite is an ergonomic wrapper for using SQLite from Rust.

serde

Serde

serde_json

Serde JSON

submission

Submit e-mail through SMTP.

utils 🔒

Structs

Configuration

The configuration for the mailpot database and the mail server.

Connection

A connection to a mailpot database.

DbVal

A database entry and its primary key. Derefs to its inner type.

MailtoAddress

A mailto: value.

QueueEntry

A queue entry.

Template

A named template.

TemplateRenderContext

Helper type for Connection::send_reply_with_list_template.

Enums

ErrorKind

The kind of an error.

Queue

In-database queues of mail.

SendMail

How to send e-mail.

Traits

ChainedError

This trait is implemented on all the errors generated by the error_chain macro.

Context

Provides the context method for Result.

ResultExt

Additional methods for Result, for easy interaction with this crate.

Functions

main 🔒

run_app 🔒

Type Definitions

Result

Convenient wrapper around std::Result.