Struct mpot_gen::models::MailingList

pub struct MailingList {
    pub pk: i64,
    pub name: String,
    pub id: String,
    pub address: String,
    pub description: Option<String>,
    pub archive_url: Option<String>,
}

A mailing list entry.

Fields
pk: i64

Database primary key.

name: String

Mailing list name.

id: String

Mailing list ID (what appears in the subject tag, e.g. [mailing-list] New post!).

address: String

Mailing list e-mail address.

description: Option<String>

Mailing list description.

archive_url: Option<String>

Mailing list archive URL.

Implementations

impl MailingList

pub fn display_name(&self) -> String

Mailing list display name (e.g. list name <list_address@example.com>).

pub fn request_subaddr(&self) -> String

Request subaddress.

pub fn id_header(&self) -> String

Value of List-Id header.

See RFC2919 Section 3: https://www.rfc-editor.org/rfc/rfc2919

pub fn help_header(&self) -> Option<String>

Value of List-Help header.

See RFC2369 Section 3.1: https://www.rfc-editor.org/rfc/rfc2369#section-3.1

pub fn post_header(&self, policy: Option<&PostPolicy>) -> Option<String>

Value of List-Post header.

See RFC2369 Section 3.4: https://www.rfc-editor.org/rfc/rfc2369#section-3.4

pub fn unsubscribe_header( &self, policy: Option<&SubscriptionPolicy> ) -> Option<String>

Value of List-Unsubscribe header.

See RFC2369 Section 3.2: https://www.rfc-editor.org/rfc/rfc2369#section-3.2

pub fn subscribe_header( &self, policy: Option<&SubscriptionPolicy> ) -> Option<String>

Value of List-Subscribe header.

See RFC2369 Section 3.3: https://www.rfc-editor.org/rfc/rfc2369#section-3.3

pub fn archive_header(&self) -> Option<String>

Value of List-Archive header.

See RFC2369 Section 3.6: https://www.rfc-editor.org/rfc/rfc2369#section-3.6

pub fn address(&self) -> Address

List address as a melib::Address

pub fn unsubscription_mailto(&self) -> MailtoAddress

List unsubscribe action as a MailtoAddress.

pub fn subscription_mailto(&self) -> MailtoAddress

List subscribe action as a MailtoAddress.

pub fn owner_mailto(&self) -> MailtoAddress

List owner as a MailtoAddress.

pub fn archive_url(&self) -> Option<&str>

List archive url value.

pub fn insert_headers( &self, draft: &mut Draft, post_policy: Option<&PostPolicy>, subscription_policy: Option<&SubscriptionPolicy> )

Insert all available list headers.

pub fn generate_help_email( &self, post_policy: Option<&PostPolicy>, subscription_policy: Option<&SubscriptionPolicy> ) -> String

Generate help e-mail body containing information on how to subscribe, unsubscribe, post and how to contact the list owners.